Rename App to Bot and move to core directory
